config: Split out tls_config implementation
Created tls_config.cc to hold implementation of TLS config
