+++ src/v/config/tls_config.h
@@ -12,21 +12,14 @@
 #pragma once
 
 #include "base/seastarx.h"
-#include "config/convert.h"
-#include "utils/to_string.h"
 
-#include <seastar/core/do_with.hh>
 #include <seastar/core/future.hh>
-#include <seastar/core/shared_ptr.hh>
 #include <seastar/core/sstring.hh>
-#include <seastar/net/ip.hh>
 #include <seastar/net/tls.hh>
 
-#include <boost/filesystem.hpp>
 #include <yaml-cpp/yaml.h>
 
 #include <optional>
-#include <utility>
 
 namespace config {
 
@@ -38,12 +31,7 @@ struct key_cert {
         return key_file == rhs.key_file && cert_file == rhs.cert_file;
     }
 
-    friend std::ostream& operator<<(std::ostream& o, const key_cert& c) {
-        o << "{ "
-          << "key_file: " << c.key_file << " "
-          << "cert_file: " << c.cert_file << " }";
-        return o;
-    }
+    friend std::ostream& operator<<(std::ostream& o, const key_cert& c);
 };
 
 class tls_config {
@@ -80,75 +68,18 @@ public:
     bool get_require_client_auth() const { return _require_client_auth; }
 
     ss::future<std::optional<ss::tls::credentials_builder>>
-    get_credentials_builder() const& {
-        if (_enabled) {
-            return ss::do_with(
-              ss::tls::credentials_builder{},
-              [this](ss::tls::credentials_builder& builder) {
-                  builder.set_priority_string("PERFORMANCE:%SERVER_PRECEDENCE");
-                  builder.set_dh_level(ss::tls::dh_params::level::MEDIUM);
-                  if (_require_client_auth) {
-                      builder.set_client_auth(ss::tls::client_auth::REQUIRE);
-                  }
 
-                  auto f = _truststore_file ? builder.set_x509_trust_file(
-                             *_truststore_file, ss::tls::x509_crt_format::PEM)
-                                            : builder.set_system_trust();
-
-                  if (_crl_file) {
-                      f = f.then([this, &builder] {
-                          return builder.set_x509_crl_file(
-                            *_crl_file, ss::tls::x509_crt_format::PEM);
-                      });
-                  }
-
-                  if (_key_cert) {
-                      f = f.then([this, &builder] {
-                          return builder.set_x509_key_file(
-                            (*_key_cert).cert_file,
-                            (*_key_cert).key_file,
-                            ss::tls::x509_crt_format::PEM);
-                      });
-                  }
-
-                  return f.then([&builder]() {
-                      return std::make_optional(std::move(builder));
-                  });
-              });
-        }
-
-        return ss::make_ready_future<
-          std::optional<ss::tls::credentials_builder>>(std::nullopt);
-    }
+    get_credentials_builder() const&;
 
     ss::future<std::optional<ss::tls::credentials_builder>>
-    get_credentials_builder() && {
-        auto ptr = ss::make_lw_shared(std::move(*this));
-        return ptr->get_credentials_builder().finally([ptr] {});
-    }
+    get_credentials_builder() &&;
 
-    static std::optional<ss::sstring> validate(const tls_config& c) {
-        if (c.get_require_client_auth() && !c.get_truststore_file()) {
-            return "Trust store is required when client authentication is "
-                   "enabled";
-        }
-
-        return std::nullopt;
-    }
+    static std::optional<ss::sstring> validate(const tls_config& c);
 
     bool operator==(const tls_config& rhs) const = default;
 
     friend std::ostream&
-    operator<<(std::ostream& o, const config::tls_config& c) {
-        o << "{ "
-          << "enabled: " << c.is_enabled() << " "
-          << "key/cert files: " << c.get_key_cert_files() << " "
-          << "ca file: " << c.get_truststore_file() << " "
-          << "crl file: " << c.get_crl_file() << " "
-          << "client_auth_required: " << c.get_require_client_auth() << ""
-          << " }";
-        return o;
-    }
+    operator<<(std::ostream& o, const config::tls_config& c);
 
 private:
     bool _enabled{false};
@@ -162,78 +93,14 @@ private:
 
 namespace YAML {
 
-inline ss::sstring to_absolute(const ss::sstring& path) {
-    namespace fs = std::filesystem;
-    if (path.empty()) {
-        return path;
-    }
-    return fs::absolute(fs::path(path)).native();
-}
-
-inline std::optional<ss::sstring>
-to_absolute(const std::optional<ss::sstring>& path) {
-    if (path) {
-        return to_absolute(*path);
-    }
-    return std::nullopt;
-}
-
 template<>
 struct convert<config::tls_config> {
-    static Node encode(const config::tls_config& rhs) {
-        Node node;
-
-        node["enabled"] = rhs.is_enabled();
-        node["require_client_auth"] = rhs.get_require_client_auth();
-
-        if (rhs.get_key_cert_files()) {
-            node["cert_file"] = (*rhs.get_key_cert_files()).key_file;
-            node["key_file"] = (*rhs.get_key_cert_files()).cert_file;
-        }
-
-        if (rhs.get_truststore_file()) {
-            node["truststore_file"] = *rhs.get_truststore_file();
-        }
-
-        return node;
-    }
+    static Node encode(const config::tls_config& rhs);
 
     static std::optional<ss::sstring>
-    read_optional(const Node& node, const ss::sstring& key) {
-        if (node[key]) {
-            return node[key].as<ss::sstring>();
-        }
-        return std::nullopt;
-    }
+    read_optional(const Node& node, const ss::sstring& key);
 
-    static bool decode(const Node& node, config::tls_config& rhs) {
-        // either both true or both false
-        if (
-          static_cast<bool>(node["key_file"])
-          ^ static_cast<bool>(node["cert_file"])) {
-            return false;
-        }
-        auto enabled = node["enabled"] && node["enabled"].as<bool>();
-        if (!enabled) {
-            rhs = config::tls_config(
-              false, std::nullopt, std::nullopt, std::nullopt, false);
-        } else {
-            auto key_cert
-              = node["key_file"]
-                  ? std::make_optional<config::key_cert>(config::key_cert{
-                    to_absolute(node["key_file"].as<ss::sstring>()),
-                    to_absolute(node["cert_file"].as<ss::sstring>())})
-                  : std::nullopt;
-            rhs = config::tls_config(
-              enabled,
-              key_cert,
-              to_absolute(read_optional(node, "truststore_file")),
-              to_absolute(read_optional(node, "crl_file")),
-              node["require_client_auth"]
-                && node["require_client_auth"].as<bool>());
-        }
-        return true;
-    }
+    static bool decode(const Node& node, config::tls_config& rhs);
 };
 
 } // namespace YAML
